requirements.txt
streamlit
langchain
langchain-community
langchain-openai
langchain-google-genai
pypdf
tiktoken
chromadb